Correccion tipo de dato precio

+++ src/test/java/vehiculos/Automovil.java
@@ -5,7 +5,7 @@ public class Automovil extends Vehiculo{
 	private static int cantidadAutomoviles;
 	
 	
-	public Automovil (String placa,String nombre, long precio, int peso,Fabricante fabricante, int puestos){
+	public Automovil (String placa,String nombre, int precio, int peso,Fabricante fabricante, int puestos){
 		super(placa,4,100,nombre,precio,peso,"FWD",fabricante);
 		this.setPuestos(puestos);
 		Automovil.cantidadAutomoviles++;
